feat: improve base layout accessibility and responsive images
Add viewport and default description meta tags, introduce skip link and navigation enhancements, and provide responsive image shortcode with WebP and JPEG support.
