+++ src/_includes/base.njk
@@ -1,7 +1,9 @@
 <!DOCTYPE html>
-<html lang="{{ page.lang }}">
+<html lang="{{ lang }}">
   <head>
     <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="description" content="{{ description or site.description }}">
     <link rel="stylesheet" href="/assets/main.css">
     <title>{{ title }}</title>
     {% if noindex %}
@@ -9,12 +11,14 @@
     {% endif %}
   </head>
   <body class="bg-white text-slate-800">
-    <a class="skip-link sr-only focus:not-sr-only focus:absolute focus:top-0 focus:left-0 focus:p-2 focus:bg-white" href="#main">Aller au contenu</a>
-    <nav aria-label="Main navigation">
-      <a href="{{ '/' | i18nPath(page.lang) }}">{{ 'home' | t(page.lang) }}</a>
-      <a href="{{ '/contact/' | i18nPath(page.lang) }}">{{ 'contact' | t(page.lang) }}</a>
+    <a href="#main" class="skip-link">Skip to content</a>
+    {% set homeUrl = '/' | i18nPath(lang) %}
+    {% set contactUrl = '/contact/' | i18nPath(lang) %}
+    <nav aria-label="Navigation principale" class="container">
+      <a href="{{ homeUrl }}"{% if page.url == homeUrl %} aria-current="page"{% endif %}>{{ 'home' | t(lang) }}</a>
+      <a href="{{ contactUrl }}"{% if page.url == contactUrl %} aria-current="page"{% endif %}>{{ 'contact' | t(lang) }}</a>
     </nav>
-    <main id="main">
+    <main id="main" class="container">
       {% block content %}{% endblock %}
     </main>
   </body>
