+++ src/assets/css/main.css
@@ -7,4 +7,25 @@
   button:focus-visible {
     @apply outline-none ring-2 ring-offset-2 ring-blue-600;
   }
+
+  .skip-link {
+    @apply sr-only focus:not-sr-only focus:absolute focus:top-0 focus:left-0 focus:p-4 focus:bg-white;
+  }
+
+  .container {
+    @apply max-w-3xl mx-auto p-4;
+  }
+
+  main p {
+    @apply mb-4;
+  }
+
+  main h1,
+  main h2,
+  main h3,
+  main h4,
+  main h5,
+  main h6 {
+    @apply mt-8 mb-4;
+  }
 }
